--- src/main/java/org/example/backedkickunity/config/domain/JpaConfig.java
@@ -1,10 +1,0 @@
-package org.example.backedkickunity.config.domain;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaConfig {
-
-}
