--- src/main/java/org/example/backedkickunity/controller/dto/JoinRequest.java
@@ -1,18 +1,0 @@
-package org.example.backedkickunity.controller.dto;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class JoinRequest {
-
-    private String memberId;
-    private String password;
-    private String email;
-    private String name;
-
-}
