--- src/main/java/org/example/backedkickunity/domain/Category.java
@@ -1,7 +1,0 @@
-package org.example.backedkickunity.domain;
-
-public enum Category {
-    MATCHING,
-    MEMBER_RECRUITING,
-    ETC_RECRUITING
-}
