--- src/main/java/org/example/backedkickunity/domain/member/Role.java
@@ -1,7 +1,0 @@
-package org.example.backedkickunity.domain.member;
-
-public enum Role {
-
-    ADMIN,
-    USER
-}
