--- src/main/java/org/example/backedkickunity/exception/BaseExceptionType.java
@@ -1,14 +1,0 @@
-package org.example.backedkickunity.exception;
-
-import org.springframework.http.HttpStatus;
-
-public interface BaseExceptionType {
-
-
-    HttpStatus getHttpStatus();
-
-    String getErrorCode();
-
-    String getErrorMessage();
-
-}
