--- src/main/java/org/example/backedkickunity/exception/ExceptionControllerAdvice.java
@@ -1,13 +1,0 @@
-package org.example.backedkickunity.exception;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class ExceptionControllerAdvice {
-
-}
-
