--- src/main/java/org/example/backedkickunity/repository/BoardRepository.java
@@ -1,7 +1,0 @@
-package org.example.backedkickunity.repository;
-
-import org.example.backedkickunity.domain.Board;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BoardRepository extends JpaRepository<Board, Long> {
-}
