--- src/main/java/org/example/backedkickunity/repository/MemberRepository.java
@@ -1,20 +1,0 @@
-package org.example.backedkickunity.repository;
-
-import org.example.backedkickunity.domain.member.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface MemberRepository extends JpaRepository<Member, Long> {
-
-    Optional<Member> findByMemberIdAndPassword(String memberId, String password);
-
-    Optional<Member> findByMemberId(String memberId);
-
-    boolean existsByMemberId(String memberId);
-
-    boolean existsByEmail(String email);
-
-}
