--- src/main/java/org/example/backedkickunity/repository/TeamRepository.java
@@ -1,7 +1,0 @@
-package org.example.backedkickunity.repository;
-
-import org.example.backedkickunity.domain.Team;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TeamRepository extends JpaRepository<Team, Long> {
-}
