+++ src/main/java/org/example/backendkickunity/config/domain/JpaConfig.java
@@ -1,0 +1,10 @@
+package org.example.backendkickunity.config.domain;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaConfig {
+
+}
