+++ src/main/java/org/example/backendkickunity/controller/dto/JoinRequest.java
@@ -1,0 +1,18 @@
+package org.example.backendkickunity.controller.dto;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public class JoinRequest {
+
+    private String memberId;
+    private String password;
+    private String email;
+    private String name;
+
+}
