+++ src/main/java/org/example/backendkickunity/controller/dto/LoginRequest.java
@@ -1,0 +1,11 @@
+package org.example.backendkickunity.controller.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class LoginRequest {
+    private String memberId;
+    private String password;
+}
