+++ src/main/java/org/example/backendkickunity/domain/Board.java
@@ -1,0 +1,17 @@
+package org.example.backendkickunity.domain;
+
+import jakarta.persistence.*;
+
+@Entity
+public class Board {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "board_id", updatable = false)
+    private Long id;
+
+    @Column(name = "title", nullable = false)
+    String title;
+
+    Category category;
+}
