+++ src/main/java/org/example/backendkickunity/domain/Category.java
@@ -1,0 +1,7 @@
+package org.example.backendkickunity.domain;
+
+public enum Category {
+    MATCHING,
+    MEMBER_RECRUITING,
+    ETC_RECRUITING
+}
