+++ src/main/java/org/example/backendkickunity/domain/member/Role.java
@@ -1,0 +1,7 @@
+package org.example.backendkickunity.domain.member;
+
+public enum Role {
+
+    ADMIN,
+    USER
+}
