+++ src/main/java/org/example/backendkickunity/exception/BaseExceptionType.java
@@ -1,0 +1,13 @@
+package org.example.backendkickunity.exception;
+
+import org.springframework.http.HttpStatus;
+
+public interface BaseExceptionType {
+    HttpStatus getHttpStatus();
+    String getErrorCode();
+    String getErrorMessage();
+}
+
+
+
+
