+++ src/main/java/org/example/backendkickunity/exception/ExceptionControllerAdvice.java
@@ -1,0 +1,9 @@
+package org.example.backendkickunity.exception;
+
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+public class ExceptionControllerAdvice {
+
+}
+
