+++ src/main/java/org/example/backendkickunity/exception/MemberException.java
@@ -1,0 +1,15 @@
+package org.example.backendkickunity.exception;
+
+public class MemberException extends BaseException {
+
+    private BaseExceptionType exceptionType;
+
+    public MemberException(BaseExceptionType exceptionType) {
+        this.exceptionType = exceptionType;
+    }
+
+    @Override
+    public BaseExceptionType getExceptionType() {
+        return exceptionType;
+    }
+}
