--- src/test/java/org/example/backedkickunity/BackedKickUnityApplicationTests.java
@@ -1,13 +1,0 @@
-package org.example.backedkickunity;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BackedKickUnityApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
