+++ src/test/java/org/example/backendkickunity/BackendKickUnityApplicationTests.java
@@ -1,0 +1,13 @@
+package org.example.backendkickunity;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BackendKickUnityApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
