+++ src/test/java/org/example/backendkickunity/service/MemberServiceTest.java
@@ -1,0 +1,15 @@
+package org.example.backendkickunity.service;
+
+import org.junit.jupiter.api.Test;
+
+class MemberServiceTest {
+
+    @Test
+    void join() {
+
+    }
+
+    @Test
+    void login() {
+    }
+}
